Two fixes to failed internal page split error handling.
After we've sucessfully split a leaf page and are climbing the tree
propagating the split upwards, we can't back everything out on failure,
the leaf page split is complete. Ignore errors other than WT_PANIC from
attempting to split higher-level tree pages, which means all internal
node page-type split functions have to clean up after themselves or panic.
This was always true, but now the code enforces it.

If splitting an internal page failed while allocating the new internal
pages for the split page's parent, we free WT_REFs that are still in use,
and nothing good happens after that.