WT-4913 Fix the Windows CRC32 on blocks that aren't 8B aligned and/or multiples of 8B (#4758)
Fix the Windows x86 install checksum. Create an alternate checksum routine we can use as a fallback if the fixed checksum fails.

Change the block manager read functions and the logging read functions to fallback to the alternate checksum if the standard checksum function fails.

Add two WiredTiger test programs to the Windows build, wt2695_checksum and wt4117_checksum.