WT-4930 review and fix likely coding errors of error returns after jump to error
* The variable ret can't be non-zero, return 0 for clarity.

* An immediate return can leak the scratch buffer allocated for building
the "-t target" argument, replace with a jump to the err: lable.

* Don't immediately return after a jump to an err: label, it's likely
to leak resources. In this particular case, we don't leak anything
but we unlock a spinlock we don't hold.

* Don't immediately return after a jump to an err: label, it's likely
to leak resources. In this particular case, I don't see any reason
inherent in the API that we can't have allocated the list-of-files
array before the return. (My suspicion is it won't happen because
the function isn't called with that combination of parameters, but
regardless, it makes no sense to have WT_ERR_MSG() and return()
calls in pairs.)