WT-4961 Checkpoints with cache overflow must keep history for reads.
When cache overflow has been active, checkpoints may need to pull pages
into cache in order to rewrite the correct versions.  These pages must
be evicted as soon as possible to avoid checkpoint filling the cache
with pages that other threads cannot evict (until checkpoint has moved
on to another tree).

This leads to the unusual case of attempting eviction of pages while
writing disk images that are correct as of the checkpoint's transaction.
In other words, __wt_txn_visible is used to select the versions of
updates written to the disk image.  However, it is not okay in this case
to discard history because everything happens to be visible to a
checkpoint: snapshot readers may need an earlier version (back to the
oldest timestamp).