WT-4460 Optimize for in-order, non-overlapping modifications (#4798)
* Allocate a WT_SESSION and pass it around for the calls that require one.
Dump failing strings during failure even if DEBUG not defined.
Dump the original string during failure.
Report the first mismatching byte during failure.

* Add an optimization path for modifications applied in sorted order where
none of the changes overlap.

* modify.c(429): warning C4267: 'function': conversion from 'size_t' to
'int', possible loss of data

* When copying the remainder of the source buffer, we should pick up
the trailing nul byte in the 'S' format, there's no need to append
an additional byte on the end.

* Simplify a comparison: once we know WT_MODIFY.size and WT_MODIFY.data.size
are the same, we only need to check one of them for overflow.

* Whitespace

* WT-4460 Fast path for in-place modifications. (#4714)

* Fast path for in-place modifications.

Avoid memory allocation in common paths (both for the buffer and for the array of modifications).  Always pack modifies before applying: that is the common case (during reads), so don't complicate the implementation by supporting a separate implementation for arrays of WT_MODIFY structs. During WT_CURSOR::modify, we're going to pack them anyway in most cases.

Add macros to walk packed modifies (forward and reverse) rather than repeating the unpack logic in multiple places.

While in the area, switch to using value->mem for writes and value->data for reads.  Remove redundant session handle from methods that take a cursor.

* Clear the key sizes, it makes it more obvious when the key is being set.

* Switch out __wt_buf_grow() for __wt_buf_grow() so WT_CURSOR.data is reset
to WT_CURSOR.mem. This is actually a bug in the previous version, which
we didn't see because all size calculations were based on offsets from the
WT_CURSOR.data value, not WT_CURSOR.mem.

* The buffer may reference an overflow item, the data may not start at the
start of the buffer's memory and we have to correct for that both when
applying modifications and when allocating memory for the modifications.