WT-5063 Returning proper error for cursor modify operation
Cursor modify is not supported for many cursor types that are
suppored in the wiredtiger. It is supported only for ds, file,
lsm and simple tables. The simple tables are internally mapped into
file cursors. Marking the rest of the cursor modify operations
as not supported to return a proper error message.

And also modify operations is supported for particular value format
types, so a new notsup function is added that checks for value format
and provide a proper error to those the modify operation is supported.