WT-4935 Add a perf test to find out the wiredtiger_calc_modify overhead (#4803)
WT-4935 Add a perf test to find out the wiredtiger_calc_modify overhead

The wiredtiger_calc_modify function is used to find out the modify vector
based on the old and new record values. This function may fail when the
changed modifications of the new record reach the allowed limit of
modification for the MODIFY operation, instead of UPDATE.

This perf test is to find out the extra overhead because of
wiredtiger_calc_modify function where it fails to find out the modify
vectors because of the reach of the limits configured for the MODIFY operation
and at the end, the operations are carried out as a normal update.

To achieve the above, a new modify workload is the new workload that is
added to wtperf to use cursor modify function instead update to update a record.
In case if the modifications are more than configured (modification is less
than 10% of the total size of the record and less than 16 modifications in the record),
the wiredtiger_calc_modify function fails to produce modify vectors,
in that such cursor update function is executed instead of modify.

For scenarios where the value size is less than 160 bytes, there is no
distribution of modifications that can occur. only the limited set of
modifications according to the test will be done.