Revert part of the original change, dd996910f9eb64d2ced057e5efbfe397b3fbe133, specifically the part of that change to mark a page clean if it was entirely visible to the evicting thread, which is in some cases the checkpoint thread.
This has a couple of problems: first, if the page has no entries that need to be written
(r->entries == 0), and it has lookaside entries (r->supd_next != 0), the page will end up as
clean with a WT_PM_REC_REPLACE result, potentially without ever being written, in which case
when the checkpoint thread tries to store the child's disk address in the parent, we fail. (This
is the failure seen in the ticket and general testing.) We could correct for that by changing
the page from WT_PM_REC_REPLACE to WT_PM_REC_EMPTY in that case (so checkpoint doesn't look
for a disk address), but that doesn't solve all of the problems: the check in eviction for
the checkpoint thread setting a page to be clean without global visibility fires, blocking the
checkpoint's eviction of the page, and what's worse, as the page is left in a clean state, it
could subsequently be evicted because it's clean without noticing there are lookaside records
required by older readers.

Revert the visibility check to leave the page dirty if it's not stable, as before. I don't
think it's possible for reconciliation to mark a page clean unless it's stable, so I've turned
the check in __evict_review() into an assert.