WT-5080 Add new workload config - distribute_modifications
The newly added workload configuration parameter can be used for
both modify/update operations where the new value record is formed
with a set of modifications distributed evenly across the whole
old record.

Along with the above, corrected the modify workload plots to run
in single thread instead of multiple for a stable result and also
added a normal update with similar configurations for comparison.