WT-4893 Fix a race between internal page child-page eviction checks and cursors in the tree(#4749) (#4857)
When checking if an internal page's children will block the internal page's
eviction, there may be cursors in the tree walking the list of child pages.
The parent is locked, so all we care about is cursors already in the child
pages, no thread can enter them. Any cursor moving through the child pages
must be hazard pointer coupling between pages, where the page on which it
currently has a hazard pointer must be in a state other than on-disk. Walk
the child list forward, then backward, to ensure we don't race with a cursor
walking in the opposite direction from our check.

Also:
* Skip work in __wt_hazard_check() and __wt_hazard_check_assert()
if we aren't maintaining hazard pointers on the tree.
* Don't clear the WT_REF history when we overwrite the WT_REF on
discard, it's useful diagnostic information.

(cherry picked from commit 82082debc5983bfc8e096bec4d812cb085cc026d)