Add the configuration option operation_timeout_ms to wiredtiger_open() and WT_SESSION.txn_begin that specifies a timeout after which the library should feel free to rollback a transaction. This includes code that optionally sets a timestamp on every API call.
Add a new test when a worker thread has been pulled into help with eviction: if the cache is
stuck, the operation_timeout_ms option is configured, and the worker thread has the oldest
pinned transaction ID, rollback the worker thread's transaction. This avoids the test program
format failing in workloads where the cache consists entirely of internal pages that cannot
be evicted because they reference child pages with lookaside data that cannot be discarded,
and the thread that potentially wants the lookaside data is trapped helping with eviction.

The reason this is linked to the operation_timeout_ms option is that the test will potentially
rollback read-only transactions, and MongoDB can't handle that.

Rework the __wt_txn_am_oldest function to avoid races between checking if the global transaction
state's WT_TXN.id is set at all, and comparing it against the running threads WT_TXN.id.

Quit inlining the __wt_txn_am_oldest function (renamed to be __wt_txn_is_blocking_old), it's
only called when eviction is stuck, there's no reason it should be performance critical.