WT-5089 Read the LAS page when the page is skewed as old
When the page is skewed to disk with oldest version, there are chances
of newer inserts that are happened on the page are not written to disk.
These inserts needs to be loaded into the page from LAS when the page
is instantiated when those inserts happened greater than or equal to the
unstable_timestamp, that is first minimum timestamp where the records
are present in the LAS. The newer inserts are read and instantiated
into the page.

For any prepared updates, those updates will be removed from disk when
they get instantiated.