WT-4715 Workloads will stall if old transaction or timestamp pinned by thread co-opted for eviction.
* In preparation for a more general timeout mechanism, rework parts of the WT-5043 changes.
Rename WT_SESSION.op_start to be WT_SESSION.op_5043_seconds and stop using WT_TRACK_TIME
to make way for a more general version. Use "WT-5043" liberally so it's obvious what to
remove in 6 months when we quit trying to reproduce this one.
Primarily, instead of doing anything at the API layer,

* Move support/time.c to support/time.i and incorporate random time functions from misc.i.
(The time functions are all small and potentially called a lot, they're good candidates
to inline.)

Add a big comment to __wt_clock(), explaining why it's OK to return ticks or nanoseconds.

* When returning seconds since the epoch, don't ignore the nanoseconds component
(there's no standard requirement they be rolled over into the seconds field in
any specific way).

Change __wt_seconds32() to call __wt_seconds() so it gets the same fix.

* Missed a check for WT_SESSION.op_5043_seconds being initialized.

* Add the configuration option operation_timeout_ms to wiredtiger_open() and WT_SESSION.txn_begin
that specifies a timeout after which the library should feel free to rollback a transaction.
This includes code that optionally sets a timestamp on every API call.

Add a new test when a worker thread has been pulled into help with eviction: if the cache is
stuck, the operation_timeout_ms option is configured, and the worker thread has the oldest
pinned transaction ID, rollback the worker thread's transaction. This avoids the test program
format failing in workloads where the cache consists entirely of internal pages that cannot
be evicted because they reference child pages with lookaside data that cannot be discarded,
and the thread that potentially wants the lookaside data is trapped helping with eviction.

The reason this is linked to the operation_timeout_ms option is that the test will potentially
rollback read-only transactions, and MongoDB can't handle that.

Rework the __wt_txn_am_oldest function to avoid races between checking if the global transaction
state's WT_TXN.id is set at all, and comparing it against the running threads WT_TXN.id.

Quit inlining the __wt_txn_am_oldest function (renamed to be __wt_txn_is_blocking_old), it's
only called when eviction is stuck, there's no reason it should be performance critical.

* Use local variable instead of repeating the assignment.

* Avoid [-Werror=maybe-uninitialized] complaints by always initializing the start time.
(It's not a particularly expensive call, shrug.)

* Run clang-format.

* Now that __wt_epoch() is in include/time.i, __wt_epoch_raw() requires default visibility.

* Clarify a comment.