WT-5146 Fix record miss for variable length columnar table
If the ref contains any LAS page, the scan searches the LAS file
for the satisfied record before returning the disk image record
to the application. During this process, cursor btree tmp buffer
is used to store the key information that is used to search the
LAS file, but using this cursor btree tmp, corrupts the memory
that is already presents in the buffer that holds the records.
Due to this, the key itself is returned to the application instead
of the actual record.

Use another temp buffer las to store the key information that is
required for scanning the LAS file instead of using the existing
tmp buffer.