WT-5146 Fix record miss for variable length columnar table (#4894)
If the ref contains any LAS page, the scan searches the LAS file
for the satisfied record before returning the disk image record
to the application. During this process, cursor btree tmp buffer
is used to store the key information for columnar tables that can
be used to search the LAS file, but using this cursor btree tmp,
corrupts the memory that is already presented in the buffer that
holds the records. Due to this, the key itself is returned to the
application instead of the actual record.

The fix is to use the local variable to hold the key for LAS search