WT-4681 Fix the las sweep to clean the obsolete records
The Sweep of LAS works by scanning the cursor from start of the LAS
till the maximum record that is found before the scan start. The records
with las_timestamp that is less than oldest_timestamp are considered
as obsolete. The identified obsolete records are not removed only when
the next record of the scan is satisfies the following conditions.
	- The next LAS record is from a different key.
	- The next LAS record is not of UPDATE MODIFY type.

In case if the next record is UPDATE MODIFY record type, the search
continues till the next record doesn't match the key or not of type
UPDATE MODIFY. If any one of the record in update chain is not obsolete
the removal is skipped.