WT-5141 Handle multiple modifies for the same key on lookaside eviction (#4900)
When we're performing lookaside eviction and we have multiple updates that would take up the same spot in lookaside (they have the same file, key, timestamp, transaction id combination), we have some logic to avoid clobbering the desired value.

In the case of standard updates, we want to write the latest value for that timestamp to the lookaside and ignore any others.

In the case of modifies, we want to squash them into a single standard update. So we look back to the most recent standard update and apply the relevant modifies on top of it.