WT-5118 Fix incorrect assert that there are no hazard references to discarded WT_REF
Because hazard pointers are acquired before the WT_REF state is checked, it's possible a thread of control running through an old page index has a hazard pointer on the WT_REF we're deleting. Remove the assertion that no hazard pointers are currently held on the WT_REF being discarded.