WT-5167 Fix the missing pieces of page instantiation from las
Update various memory allocations that are done during the page
instantiation for the tracking stats variables.

Set the page modify restore_state with proper value once the page
is successfully instantiated.

If LAS updates are newer than the disk image, set the flag to include
these updates in the next checkpoint.

Otherwise, if the page image contains newest versions of data than
LAS updates, check whether the page that is marked as dirtly can
be cleared under some conditions.