WT-5167 Fix the missing pieces of page instantiation from las (#4911)
Update various memory allocations that are done during the page
instantiation for the tracking stats variables.

Set the page modify structure restore_state with proper value once the page
is successfully instantiated.

If LAS updates are newer than the disk image, set the flag to include
these updates in the next checkpoint.

Otherwise, if the page image contains the newest versions of data than
LAS updates, check whether the page that is marked as dirty can
be cleared under some conditions.

Fix the scan-build uninitialized variable usage warning

Correct the statistics name in the test. The statistics name is changed from cache_read_lookaside to cache_lookaside_read in wt-4685.