WT-5158 Remove assertion in __wt_find_lookaside_upd for cursor comparison (#4912)
This assertion is incorrect. The WT_CURSOR::next code path does not touch the compare value however it may need to read updates from lookaside. So if we start with a non-zero comparison for whatever reason (earlier failed search is one example), and then attempt a "next" call, we'll hit this.