WT-5145 Fix the race condition in accessing pinned_timestamp and connection state
The __wt_txn_visible_all function refers both has_pinned_timestamp
and connections state to decide whether the data is visible or not.
But during the connection close flow, the two variables are set in
two different functions and there is a very chance of connection swtich
can leads to wrong usage.

Set the connection state as WT_CONN_CLOSING as the first step of
connection close operation, as this can avoid the race condition
and fix the assert problem.