WT-4681 Fix the las sweep to clean the obsolete records (#4903)
The Sweep of LAS works by scanning the cursor from the start of the LAS
till the maximum record that is found before the scan start. The records
with las_timestamp that is less than oldest_timestamp are considered
as obsolete. The identified obsolete records are not removed only when
the next record of the scan is satisfied with the following conditions.
	- The next LAS record is from a different key.
	- The next LAS record is not of UPDATE MODIFY type.

In case if the next record is UPDATE MODIFY record type, the search
continues till the next record doesn't match the key or not of type
UPDATE MODIFY. If anyone of the record in the update chain is not obsolete
the removal is skipped.

There are chances of additional inserts at key boundary, so there are chances of reading a different record whenever we try to move to the previous record to remove. To avoid such scenarios, the first previous record is compared with the saved key whenever there is a key change.