WT-5191 LAS sweep to consider last checkpoint time
Sometimes it is possible that the obsolete records in
lookaside are still not may be written to disk image. To avoid
removal of those records, consider the last_ckpt_timestamp also
into consideration before removing those records.

Checkpoint will ensure read all the pages in the tree that has
lookaside content for checkpoint, this ensures that proper version
of the lookaside records are written to disk image, so considering
the last checkpoint time also in obsolete record removal will avoid
of removing any records that don't have disk image.