Handle cases where no timestamps and contains a previous ondisk value
Whenever there are no timestamps enabled in the system, the obsolete
entries are identified using the transaction id mechanism. Along with
that the special case added for the lookaside record that is a previous
ondisk value is only removed when there is a lookaside record for the
same key is removed.