WT-5191 Additional LAS sweep to consideration before obsoleting a record (#4928)
Sometimes it is possible that the obsolete records in
lookaside are still not may be written to the disk image. To avoid
removal of those records, consider the last_ckpt_timestamp also
into consideration before removing those records.

The checkpoint will ensure read all the pages in the tree that has
lookaside content for checkpoint, this ensures that proper version
of the lookaside records are written to the disk image, so considering
the last checkpoint time also in obsolete record removal will avoid
of removing any records that don't have disk image.

Whenever there are no timestamps enabled in the system, the obsolete
entries are identified using the transaction id mechanism. Along with
that the special case added for the lookaside record that is a globally visible
on-disk value is only removed when there is a lookaside record for the
the same key is removed.