WT-4844 Only log an informational message when a set read-timestamp is older than the oldest timestamp. (#4669)
* In some cases, MongoDB sets a read timestamp older than the oldest timestamp,
relying on WiredTiger's concurrency to detect and fail the set. In other
cases it's a bug and MongoDB wants error context to make it easier to find
those problems. Don't output an error message because that logs a MongoDB
error, use an informational message to provide the context instead.

(cherry picked from commit 2fa11c0)