WT-5186 Handle reading of LAS modifies on top of a birthmark record (#4934)
When reading a modify from the lookaside either as part of lookaside instantiation or a regular read, we need to be aware that the base update may not be in the lookaside. It's possible that the base update is either the birthmark or the on-disk value (and specifically NOT a birthmark).

There was an existing issue with the "small vector" logic used for storing modifies as part of lookaside page instantiation. This patch fixes this and includes a light refactor to consolidate this repeated logic into a WT_MODIFY_VECTOR data structure.