Add a new cursor flag that modifies WT_CURSOR.updates to search any pinned page before doing a full tree search. Currently internal only, used for LAS inserts.
Historically, we used either (WT_CURSOR_BTREE.slot != 0) or (WT_CURSOR_BTREE.compare == 0) to
determine when a leaf-only search wasn't possible because the key was outside the pinned cursor
page's name space (in other words, the search was only successful if the searched-for key was
within the leaf page's name space). I know how we got here, but it's getting messy and there are
already too many different tests of search "success": change __wt_row_search/__wt_col_search to
take an additional argument that returns if the leaf-page search wasn't possible. Reorder the
__wt_row_search/__wt_col_search arguments so all 3 of the leaf-page-search oriented arguments
are in order at the end of the argument list, triggered by a non-NULL leaf page.