WT-5196 Let the LAS sweep works under read committed isolation
LAS sweep working under read uncommitted isolation can find the
LAS records that are even not committed. The LAS sweep does the
cleanup based on the first record that is present for the key.
As the LAS sweep operates in read uncommitted mode, sometimes
even when the entire record chain for the key is not finished
insertion, the LAS sweep may clean the part of the update chain
as it finds them as obsolete. This leads to missing update chain
records and that leads to record miss.

To solve this problem, let the LAS sweep to work under read
committed isolation level, that leads to non visibility of records
from a running transaction.