WT-5142 Don't create huge root pages when rebalancing or bulk-loading objects with overflow keys
* Overflow items in an internal page can lead to (pinned) root pages larger than the maximum size expected for in-memory objects and a resulting eviction churn.

Treat rebalance handles like bulk load handles and force a close when the handle is discarded. Bulk loads and rebalanced trees are special because they may have huge root pages in memory, and we need to push those pages out of the cache. The only way to do that is to close the handle.

Drill a hole into the boundary crossing routines that support the caller forcing a split. This is currently only used when reconciling row-store internal pages, where overflow keys can cause the in-memory size of the disk image we're creating to be larger than the maximum size supported by the configuration.

Don't test WT_PAGE.WT_PAGE_OVERFLOW_KEYS when reconciling a page. It's only set when reading a page from disk, and there are enough ways we create pages in memory that I think it's risky to depend on it when reconciling a random page. I don't think this fixes any existing bugs, it's a safety measure for future changes.

Move reconciliation's buffer image chunk size into the WT_RECONCILE structure: previously, it was calculated and then used when initializing the first chunk, and when needed again, taken from the chunk known to be initialized, and that's fragile/confusing.

Split __wt_rec_need_split() into two functions, __wt_rec_need_split() and __rec_need_split_leaf(), and move Btree row-leaf specific tess from the former into the latter. This simplifies the boundary crossing macros that we need (WT_CHECK_CROSSING_BND goes away).

* rec_row.c:505:17: warning: Value stored to 'key_onpage_ovfl' is never read

* Revert a chunk of the original change, it changes eviction characteristics and that's not
something I want to do.

* Don't include the size of the normal key/value pair, because then we're arguing with the
already existing compression configuration.

* Rename page_image to key_overflow_size, the variable only tracks the bytes allocated to
overflow keys.