WT-5126 Incorrect assertion regarding prepared update resolution
Remove the commit/rollback assertion that the number of prepared updates, visited based on
the transaction modify array's list of updates, is the same as the number of prepared updates
resolved by processing the transaction modify array's list of updates. The problematic case
is as follows: first a transaction deletes a page as part of a fast truncate, then the deleted
page is instantiated (creating individual tombstone updates for each deleted key/value pair),
then the same transaction subsequently updates a key/value item on the deleted page. In this
scenario, the update for the original delete does not appear in the transaction's modify list,
the subsequent update does appear in the transaction's modify list, and both updates appear in
the on-page list of updates for the key/value item.
    When preparing a transaction, for each update we check if the transaction made other updates
to the same key/value item, and if that is the case, we set a flag to ignore all but one of the
updates when resolving the transaction. (This is because it's a lot of work to resolve an update
in a prepared transaction, and it's cheaper to resolve all of the updates to each key/value pair
at the same time.) However, in the scenario described above, we will find more updates in the
update list than there are updates in the transaction's modify list because we find the fast
delete update when reviewing the list of on-page updates based on the transaction modify list.
For this reason, we cannot assert that the number of updates we find in the transaction's
modify list is equivalent to the number of updates actually resolved when the on-page updates
are reviewed.

Remove the txn_prepared_updates_resolved statistic -- it's the same as the
txn_prepared_updates_count statistic, and all it can tell us if there's a
lag between preparing the updates and their eventual resolution, and the
existing txn_prepare_XXX statistics are going to do a better job of that.