WT-5136 Fix reading freed memory due to birthmark after uncommitted updates freed  (#4962)
Obsolete birthmark cannot be freed if it follows a set of uncommitted
updates as the uncommitted updates can be aborted and freed causing the
entire update chain being removed in this case.

Birthmark can be discarded only if there is an update that is visible to
all before it.

(cherry picked from commit bb77fe68986cc9ff26e516f05db481cbc137fab6)