WT-5264 Add delete record for every insert into history store to represent stop time pair
Every record that gets inserted into a history store must have a stop timepair.
To represent the stop timepair to automatically propagated to the cell structure,
add a delete record for every insert record with a stop time pair as the commit time pair.

During reconcilation of a history page, these two records must be reconciled into
a single record with that have both start and stop time pairs.

History store records are special as they store the history of an actual record.
we need to retain the transaction and timestamp of those records as it was done by
the original transaction and timestamp, so make sure that the updates doesn't get
modified or validated for the history store transaction id during commit/rollback.