WT-5082 application threads are tasked with eviction even when pinning the oldest transaction ID (#4996)
* A read-only, read-committed transaction will have a snapshot which may be the oldest transaction
ID in the system, but won't have a matching transaction ID. When looking to match the oldest
transaction ID in the system, check the older of the transaction ID or the snapshot start point.

* Revert "A read-only, read-committed transaction will have a snapshot which may be the oldest transaction"

This reverts commit be2946098b1557b59653c308a2e07d95252cf7f0.

* WT-4715 incorrectly set the operation timeout values, operations that didn't
explicitly configure transactions didn't default to any WT_CONNECTION
configured timeout. Rework that change: starting an explicit transaction
sets/clears a WT_TXN value when the transaction is configured/reset, and the
per-operation API calls set/clear timeouts based on the WT_TXN timeout, or,
if that value isn't set, the WT_CONNECTION timeout.

* We have a function to check if the current transaction ID is the oldest transaction ID in
the system and another function to check if the current transaction's snapshot minimum is the
oldest snapshot minimum in the system. This isn't right: if a transaction's snapshot is the
oldest snapshot in the system, it's still pointless to roll it back if another transaction's
ID is the oldest transaction ID in the system. Merge the two functions into a single check.

* On API_XXX_INIT, clear the timer values if we're not starting a timer, that way we're less
likely to introduce a bug if the API_XXX_END macros ever change to not line up the way they
do now.

* Instead of reviewing active transactions/sessions for the lowest transaction ID in the system,
compare against the maintained txn_global->oldest_id value.