WT-5267 Store stop time pair in history store (#4999)
* WT-5267 Store stop time pair in history store

Stop time pair is required for every record in the history
store to represent the end of visibility of each record in
the history store. Store it in the key of the every history
record.

* Use durable timestamp instead of commit timestamp as stop timestamp

* Update uses of las cursor key to use new format

The key format is changed to include stop time pair, so
all the locations that are using the las cursor are updated
to use the new format. Currently the retrieved stop time pair
values are not used widely.

* Avoid wrapping variable declarations into multi lines

* Use of WT_TIME_PAIR structure whereever possible

* Revert to use commit timestamp instead of durable timestamp

Commit timestamp of next record represents the end of visibility
of the current record.

* comit => commit correction