Support wiredtiger builds with "attach" functionality. This requires detaching each format run from the shell script (otherwise the script couldn't exit without killing the process), and tracking running processes differently because once a process stops to be attached, it's no longer counted as a running job.
Acquire the wt binary path if we're running the smoke tests, not just abort/recovery tests.

Rename "stop" as "quit", to match "force_quit".

Add "--" as an argument to separate format.sh arguments from format arguments so it's possible
to pass normal arguments (for example, "-l") to format.

Whitespace cleanups.