Fix two cases where successful pinned page searches can result in attempting to modify an object before starting the necessary implicit transaction: WT_SESSION.truncate and WT_SESSION.update (with the WT_CURSTD_UPDATE_LOCAL flag set).
We haven't seen the failure in truncate before because a "read" of any page normally
results in starting the transaction, it's only pinned pages that avoid that operation,
and it's relatively unlikely both the start/stop cursors involved in a range truncate
operation are the result of successful pinned page searches.

We haven't seen the failure in update because it's currently only possible when updating
the lookaside table, which uses explicit transactions.